pyproject.toml
[tool.poetry]
name = "jamhacks8"
version = "0.1.0"
description = ""
authors = ["Jason Cameron <git@jasoncameron.dev>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
google-auth = "^2.3.3"
google-auth-oauthlib = "^0.4.6"
google-auth-httplib2 = "^0.1.0"
cryptography = "^42.0.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
